requirements.txt
lime==0.2.0.1
matplotlib==3.7.5
numpy==1.24.4
pillow==10.4.0
scikit-image==0.21.0
scikit-learn==1.3.2
torch==2.4.1
torchcam==0.4.0
torchmetrics==1.5.2
torchvision==0.19.1
tqdm==4.67.1
